+++ src/Core/Dfe.Complete.Domain/Enums/ProjectsOrderBy.cs
@@ -9,6 +9,5 @@ public enum OrderProjectByField
 {
     CompletedAt,
     CreatedAt,
-    SignificantDate,
-    Id
+    SignificantDate
 }