+++ src/Core/Dfe.Complete.Domain/Enums/UserProjectsOrderBy.cs
@@ -1,0 +1,7 @@
+namespace Dfe.Complete.Domain.Enums;
+
+public enum OrderUserProjectsByField
+{
+    Id,
+    SignificantDate
+}